Improve README; Update Configuration Placeholders
Updated some of the configuration placeholders. There were some fields that should have placeholder values, but didn't. Updated the README to reflect some of the missing information like running the Blazor application or loading the configuration data into Atlas. Also changed the target version of .NET to be 6.0 rather than 7.0.